do_build.sh : Get rid of of HOME.
This was to keep git happy for builds where the user doing the build
had no homedir or git config. This was required in two scenarios:
our use of guilt that is no longer the case due to the pq class going
away, and builds run by users without git configured. The prior isn't
an issue any longer and the later is a deficiency of the builder. No
need to complicate our build scripts on account of misconfigured
builders.

OXT-94

Signed-off-by: Philip Tricca <flihp@twobit.us>