Drop usage of gene3fs.
The problem is that there are a small number of unlabeled files in the final
image produced by gene3fs, whereas building the filesystem images using the
normal tools does not exhibit this problem.  It turns out that these are
cases where there were multiple hard links to a single file (e.g. e2fsck,
fsck.ext2, fsck.ext3, fsck.ext4).  In the pseudo files database, only one of
the instances is getting its attribute recorded, and in the filesystem image
that is generated by gene3fs, multiple hard links to the same inode are not
being preserved; multiple separate inodes are being created and the other
instances are never getting an xattr set.  In contrast, mkfs.ext3 preserves
the hard links as links and correctly labels the inode.

In the old pseudo xattr patch in openxt, xattrs were fetched by device,
inode, and xattr name, so queries on the same inode will yield the same
result.  In the upstream pseudo xattr implementation, it looks like they
are fetching by file id and xattr name, where file ids are keys into the
files table, which includes pathname information in addition to device
and inode.  So it seems that a getxattr on any of the hard links that did
not have setxattr() called on it will fail.  And setfiles won't bother
relabeling a file if it already has the right label, so it won't call setxattr
on more than one of the hard links.

Arguably the upstream pseudo xattr implementation is buggy, but it appears
to follow the same model as for other attributes and it "works" when using
the upstream image creation tools, so it doesn't seem likely that they would
be interested in a "fix" there.

Before:
12867 system_u:object_r:unlabeled_t:s0 /sbin/e2fsck
12799 system_u:object_r:fsadm_exec_t:s0 /sbin/fsck.ext2
12861 system_u:object_r:unlabeled_t:s0 /sbin/fsck.ext3
12882 system_u:object_r:unlabeled_t:s0 /sbin/fsck.ext4
12801 system_u:object_r:unlabeled_t:s0 /sbin/fsck.ext4dev
(All separate inodes, and only one of them is labeled)

After:
11774 system_u:object_r:fsadm_exec_t:s0 /sbin/e2fsck
11774 system_u:object_r:fsadm_exec_t:s0 /sbin/fsck.ext2
11774 system_u:object_r:fsadm_exec_t:s0 /sbin/fsck.ext3
11774 system_u:object_r:fsadm_exec_t:s0 /sbin/fsck.ext4
11774 system_u:object_r:fsadm_exec_t:s0 /sbin/fsck.ext4dev
(All hard links to a single inode, which is labeled)

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>