do_build.sh: Remove redundant build configs.
Both PARALLEL_MAKE and BB_NUMBER_THREADS are defined in the local.conf
and then overridden by the do_build.sh script. This patch removes the
generated values from do_build and leaves the static defaults in
local.conf. We also remove the mention of OE_BB_TREADS value in .config
as it was used to generate the BB_NUMBER_THREADS value in do_build.
This value was undocumented (as noted in the config file) so it's
probably best to rely on the OE site.conf mechanism.

Signed-off-by: Philip Tricca <flihp@twobit.us>