Allow meta-selinux to follow the jethro branch.
commit 096a5ee ("Clone meta-selinux at known good rev.") switched
meta-selinux to using a specific commit because at the time,
meta-selinux had no jethro branch and changes in the master branch
broke the build.  Since that time, Phil created a jethro branch for
meta-selinux, so we can switch back to letting meta-selinux default
to following OE_BRANCH.   In this manner, we ensure that any subsequent
jethro-compatible fixes and improvements to meta-selinux will be picked up
by future OpenXT builds.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>