setup_build: Use git submodules for repos
We already configure all the OE repos with git submodules, so re-use
that configuration by default.  This will keep do_build.sh and
build-scripts more in sync.

If a $REPO_REPO is defined, that is used for checkout instead of the
submodule.  If $REPO_TAG is defined, that revision is checked out.

Bitbake is included in the submodules, so allow it to be checked out by
that means.  $BB_BRANCH is now the default branch if $BITBAKE_REPO is
defined.  $BITBAKE_TAG specifies the tree to checkout.

meta-openxt-{ocaml,haskell}-platform are OpenXT branches and not in
submodules, so they must be specified by environment variable.  Also,
they follow the OpenXT ${BRANCH} by default.

OXT-880

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
(cherry picked from commit 91dccd1de25da5c1cbcb062e7d478c3fde5a9798)