example-config: Set BB_BRANCH to specific revision
stable-8 is tracking specific revisions, so set BB_BRANCH to match the
specific revision in the git submodule.

By default, do_build will checkout the submodule.  If BITBAKE_REPO is
set and BITBAKE_TAG is empty, BB_BRANCH is used as the default.  We need
a revision to prevent checking out the branch head.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>