do_build.sh: test for bash as system shell
If /bin/sh is not symlinked to bash, some packages in the OpenXT
build will fail with non-obvious error messages. Before starting
the build, confirm that the default non-interactive shell is bash.
In Debian/Devuan, the default is dash.

Future improvement: identify and patch each offending package to
explicitly require bash in shell scripts. This would avoid imposing
the overhead of bash on all build scripts, at the cost of OpenXT
carrying patches until they could be upstreamed.

OXT-1736

Signed-off-by: Rich Persaud <rich.persaud@baesystems.com>