build-scripts: Remove installer-part2-image
control.tar.bz2 is now created directly in
xenclient-installer:do_deploy, so installer-part2-image no longer
exists.  It is generated as a byproduct of xenclient-installer-image.

Update to remove xenclient-installer-part2-image and grab the tarball as
the correct name.  With the removal of xenclient-installer-part2-image,
the artifacts previously copied there are not copied in
xenclient-installer-image.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>